Add sync-on-exit option
This adds a flag to clsync called --sync-on-exit. The goal of this flag
is to have "soft kill" do a final sync of any changes in the queue
before exiting.

This was done by added to the current STATE_PREEXIT and prepending a
check that the flag was set and then telling the context to exit on no
events. This should ensure that everything is out of the event queue one
before exiting. This hows the downside of possibly continual running
until the event queue becomes empty.

The reason that this logic was added above the `main_status_update` call
is because when it was below that statement there was a segfault.